Starting by renaming the storageHandler to jsonHandler
