+++ resources/storage/shortcuts.json
